requirements.txt
flask>=3.0.0
pybaseball>=2.2.7
python-dateutil>=2.8.2
python-dotenv>=1.0.0
tabulate>=0.9.0
